Fixes #9077: Inventory "toInt" error when parsing report with VIRTUALMEMORY = 2297795780

+++ inventory-api/src/main/scala/com/normation/inventory/domain/NodeInventory.scala
@@ -89,7 +89,7 @@ case class Process (
   , started       : Option[String]   = None
   , tty           : Option[String]   = None
   , user          : Option[String]   = None
-  , virtualMemory : Option[Int]      = None
+  , virtualMemory : Option[Double]   = None
   , description   : Option[String]   = None
 )extends NodeElement with HashcodeCaching
